requirements.txt
tqdm
dvc
dvc[gs]
pymongo[srv]==4.2.0
pandas
numpy
PyYAML
scikit-learn
python-dotenv
cryptography
mlflow
boto3
pytest==7.2.2
#tox
#flake8
gcpx
pycaret==3.0.0rc9
evidently
-e .
